requirements.txt
mlflow==2.6.0
openai==2.8.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
